Add files via upload
this project can be used to create a basic web page using HTML, DAO, loginController. the html page contains login page which by giving the correct credentials( in this case , if userid=password) terminates from the login page to home page or else remains in the login page with a message saying "invalid credentials".